Create Conditional Output and Looping in VBScript
This VBScript snippet initializes two variables (a and b) with values 1 and 2, respectively. It then checks if a is greater than or equal to 1 and if b is less than or equal to 3. If both conditions are true, it outputs the message "beta1" to the console.

Following this, the script enters a For loop where c iterates from 0 to 1. During each iteration, it outputs "a is alpha" followed by the current value of c to the console. Finally, the script terminates.